+++ src/neural_network/core/utilities.rs
@@ -1,8 +1,8 @@
 use rand::thread_rng;
 use rand_distr::{Distribution, Normal};
 
-pub fn he_initialise(dim_in: usize, dim_out: usize) -> Vec<f64> {
-    let factor = (2.0 / (dim_in as f64)).sqrt();
+pub fn he_initialise(dim_in: usize, dim_out: usize) -> Vec<f32> {
+    let factor = (2.0 / (dim_in as f32)).sqrt();
     let mut rng = thread_rng();
     let normal = Normal::new(0.0, factor).unwrap();
